+++ include/engine/trip/trip_brute_force.hpp
@@ -31,7 +31,22 @@ EdgeWeight ReturnDistance(const util::DistTableWrapper<EdgeWeight> &dist_table,
     std::size_t i = 0;
     while (i < location_order.size() && (route_dist < min_route_dist))
     {
-        route_dist += dist_table(location_order[i], location_order[(i + 1) % component_size]);
+
+        auto edge_weight = dist_table(location_order[i], location_order[(i + 1) % component_size]);
+
+        // If the edge_weight is very large (INVALID_EDGE_WEIGHT) then the algorithm will not choose
+        // this edge in final minimal path. So instead of computing all the permutations after this
+        // large edge, discard this edge right here and don't consider the path after this edge.
+        if (edge_weight == INVALID_EDGE_WEIGHT)
+        {
+            return INVALID_EDGE_WEIGHT;
+        }
+        else
+        {
+            route_dist += edge_weight;
+        }
+
+        // This boost assert should not be reached if TFSE table
         BOOST_ASSERT_MSG(dist_table(location_order[i], location_order[(i + 1) % component_size]) !=
                              INVALID_EDGE_WEIGHT,
                          "invalid route found");